# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Investing Algorithm API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface AnalyzeRequest {
  /** Stock ticker symbol (e.g. AAPL) */
  ticker: string;
}

export interface FundamentalScore {
  peRatio?: number | null;
  pbRatio?: number | null;
  debtToEquity?: number | null;
  currentRatio?: number | null;
  revenueGrowth?: number | null;
  epsGrowth?: number | null;
  profitMargin?: number | null;
  roe?: number | null;
  dividendYield?: number | null;
  /** 0-100 fundamental score */
  score: number;
}

export interface TechnicalScore {
  rsi?: number | null;
  macdSignal?: string | null;
  sma50Above200?: boolean | null;
  volumeTrend?: string | null;
  priceVs52WeekHigh?: number | null;
  /** 0-5 technical score */
  score: number;
}

/**
 * Tier classification
 */
export type StockAnalysisTier =
  (typeof StockAnalysisTier)[keyof typeof StockAnalysisTier];

export const StockAnalysisTier = {
  "A+": "A+",
  A: "A",
  B: "B",
  C: "C",
} as const;

export type StockAnalysisEntrySignal =
  (typeof StockAnalysisEntrySignal)[keyof typeof StockAnalysisEntrySignal];

export const StockAnalysisEntrySignal = {
  Strong_Buy: "Strong Buy",
  Buy: "Buy",
  Hold: "Hold",
  Wait: "Wait",
} as const;

export interface StockAnalysis {
  ticker: string;
  companyName?: string | null;
  currentPrice?: number | null;
  /** 0-100 overall score (70% fundamental + 30% technical) */
  overallScore: number;
  /** Tier classification */
  tier: StockAnalysisTier;
  entrySignal: StockAnalysisEntrySignal;
  /** Automatic 8% stop-loss price */
  stopLoss?: number | null;
  /** 20% profit target */
  target1?: number | null;
  /** 35% profit target */
  target2?: number | null;
  /** 50% profit target */
  target3?: number | null;
  fundamental: FundamentalScore;
  technical: TechnicalScore;
  analyzedAt: string;
  sector?: string | null;
  marketCap?: number | null;
}

export interface RecommendationsResponse {
  items: StockAnalysis[];
  total: number;
}

export interface AddPositionRequest {
  ticker: string;
  companyName?: string | null;
  shares: number;
  entryPrice: number;
  tier?: string | null;
  entrySignal?: string | null;
  stopLoss?: number | null;
  target1?: number | null;
  target2?: number | null;
  target3?: number | null;
  notes?: string | null;
}

export type UpdatePositionRequestStatus =
  | (typeof UpdatePositionRequestStatus)[keyof typeof UpdatePositionRequestStatus]
  | null;

export const UpdatePositionRequestStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface UpdatePositionRequest {
  currentPrice?: number | null;
  shares?: number | null;
  status?: UpdatePositionRequestStatus;
  exitPrice?: number | null;
  notes?: string | null;
}

export type PortfolioPositionStatus =
  (typeof PortfolioPositionStatus)[keyof typeof PortfolioPositionStatus];

export const PortfolioPositionStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface PortfolioPosition {
  id: number;
  ticker: string;
  companyName?: string | null;
  shares: number;
  entryPrice: number;
  currentPrice?: number | null;
  status: PortfolioPositionStatus;
  exitPrice?: number | null;
  tier?: string | null;
  entrySignal?: string | null;
  stopLoss?: number | null;
  target1?: number | null;
  target2?: number | null;
  target3?: number | null;
  notes?: string | null;
  openedAt: string;
  closedAt?: string | null;
  unrealizedPnl?: number | null;
  unrealizedPnlPct?: number | null;
  realizedPnl?: number | null;
}

export interface PortfolioSummary {
  totalValue: number;
  totalCost: number;
  totalUnrealizedPnl: number;
  totalUnrealizedPnlPct: number;
  totalRealizedPnl: number;
  openPositions: number;
  closedPositions: number;
  winRate?: number | null;
}

export interface PortfolioResponse {
  positions: PortfolioPosition[];
  summary: PortfolioSummary;
}

export interface AddWatchlistRequest {
  ticker: string;
  companyName?: string | null;
  notes?: string | null;
  alertPrice?: number | null;
}

export interface WatchlistItem {
  id: number;
  ticker: string;
  companyName?: string | null;
  notes?: string | null;
  alertPrice?: number | null;
  addedAt: string;
  lastAnalysis?: StockAnalysis | null;
}

export interface WatchlistResponse {
  items: WatchlistItem[];
  total: number;
}

export type GetRecommendationsParams = {
  tier?: GetRecommendationsTier;
  signal?: GetRecommendationsSignal;
  limit?: number;
};

export type GetRecommendationsTier =
  (typeof GetRecommendationsTier)[keyof typeof GetRecommendationsTier];

export const GetRecommendationsTier = {
  "A+": "A+",
  A: "A",
  B: "B",
  C: "C",
} as const;

export type GetRecommendationsSignal =
  (typeof GetRecommendationsSignal)[keyof typeof GetRecommendationsSignal];

export const GetRecommendationsSignal = {
  Strong_Buy: "Strong Buy",
  Buy: "Buy",
  Hold: "Hold",
  Wait: "Wait",
} as const;