# lib/api-zod/src/generated/types/addPositionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Investing Algorithm API
 * OpenAPI spec version: 0.1.0
 */

export interface AddPositionRequest {
  ticker: string;
  companyName?: string | null;
  shares: number;
  entryPrice: number;
  tier?: string | null;
  entrySignal?: string | null;
  stopLoss?: number | null;
  target1?: number | null;
  target2?: number | null;
  target3?: number | null;
  notes?: string | null;
}