# lib/api-zod/src/generated/types/addWatchlistRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Investing Algorithm API
 * OpenAPI spec version: 0.1.0
 */

export interface AddWatchlistRequest {
  ticker: string;
  companyName?: string | null;
  notes?: string | null;
  alertPrice?: number | null;
}