# lib/api-zod/src/generated/types/fundamentalScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Investing Algorithm API
 * OpenAPI spec version: 0.1.0
 */

export interface FundamentalScore {
  peRatio?: number | null;
  pbRatio?: number | null;
  debtToEquity?: number | null;
  currentRatio?: number | null;
  revenueGrowth?: number | null;
  epsGrowth?: number | null;
  profitMargin?: number | null;
  roe?: number | null;
  dividendYield?: number | null;
  /** 0-100 fundamental score */
  score: number;
}