# lib/api-zod/src/generated/types/portfolioResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Investing Algorithm API
 * OpenAPI spec version: 0.1.0
 */
import type { PortfolioPosition } from "./portfolioPosition";
import type { PortfolioSummary } from "./portfolioSummary";

export interface PortfolioResponse {
  positions: PortfolioPosition[];
  summary: PortfolioSummary;
}