# lib/api-zod/src/generated/types/recommendationsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Investing Algorithm API
 * OpenAPI spec version: 0.1.0
 */
import type { StockAnalysis } from "./stockAnalysis";

export interface RecommendationsResponse {
  items: StockAnalysis[];
  total: number;
}