# lib/api-zod/src/generated/types/updatePositionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Investing Algorithm API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdatePositionRequestStatus } from "./updatePositionRequestStatus";

export interface UpdatePositionRequest {
  currentPrice?: number | null;
  shares?: number | null;
  status?: UpdatePositionRequestStatus;
  exitPrice?: number | null;
  notes?: string | null;
}