# lib/api-zod/src/generated/types/watchlistItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Investing Algorithm API
 * OpenAPI spec version: 0.1.0
 */
import type { StockAnalysis } from "./stockAnalysis";

export interface WatchlistItem {
  id: number;
  ticker: string;
  companyName?: string | null;
  notes?: string | null;
  alertPrice?: number | null;
  addedAt: Date;
  lastAnalysis?: StockAnalysis | null;
}